refactor(i18n): enhance locale management in Layout context
- Integrated browser locale detection into the Layout context for improved user experience.
- Updated locale initialization to prioritize user preferences and browser settings.
- Removed redundant browser locale function from Locale context, streamlining locale handling across the application.
